requirements.txt
Flask==2.1.0
Flask-CORS==3.0.10
pymongo[srv]==4.1.1
python-dotenv==0.19.2
gunicorn==20.1.0
Werkzeug==2.0.3